add idle=halt to F16 kernel command line for EC2
This is necessary to work around some Xen server-side shortcomings.
Without it instances will panic on launch, though only some of the time.
We believe this is a result of differences in the underlying Xen version
used in different EC2 regions and even within a given region.

An Ubuntu bug tracking some of the details:

https://bugs.launchpad.net/ubuntu/+source/linux/+bug/881076

This is allegedly addressed in the upstream kernel in 3.2 rc5.
We'll have to carry this fix as long as F16 is 3.1 based.