add locking to fix two concurrency bugs
Our downloads from warehouse and compression of images prior to EBS upload pushes both
contain concurrency bugs.  Locking ensures that only one builder will grab a given image
at a given time and that only one builder will compress that image at any given time.

Without this, horrible things happen.  (A builder uploads a gz file that is in the process
of being created or compresses an image file that is in the process of being retrieved.)